--- packages/language-service/ivy/compiler/BUILD.bazel
@@ -1,17 +1,0 @@
-load("//tools:defaults.bzl", "ts_library")
-
-package(default_visibility = ["//packages/language-service/ivy:__pkg__"])
-
-ts_library(
-    name = "compiler",
-    srcs = glob(["*.ts"]),
-    deps = [
-        "//packages/compiler-cli",
-        "//packages/compiler-cli/src/ngtsc/core",
-        "//packages/compiler-cli/src/ngtsc/file_system",
-        "//packages/compiler-cli/src/ngtsc/incremental",
-        "//packages/compiler-cli/src/ngtsc/typecheck",
-        "//packages/compiler-cli/src/ngtsc/typecheck/api",
-        "@npm//typescript",
-    ],
-)
