--- packages/language-service/ivy/compiler/README.md
@@ -1,2 +1,0 @@
-All files in this directory are temporary. This is created to simulate the final
-form of the Ivy compiler that supports language service.
