+++ packages/language-service/ivy/test/BUILD.bazel
@@ -25,7 +25,6 @@ jasmine_node_test(
     ],
     tags = [
         "ivy-only",
-        "manual",  # do not run this on CI since compiler APIs are not yet stable
     ],
     deps = [
         ":test_lib",
